+++ R/plot-functions.R
@@ -94,6 +94,13 @@ plot_db <- function(data, cluster = FALSE, legend = FALSE) {
 #'    represented as a heatmap of probabilities for 1-dimensional inputs. If
 #'    FALSE (default), the mean curve and associated Credible Interval are
 #'    displayed.
+#' @param samples A logical value indicating whether the GP prediction should be
+#'    represented as a collection of samples drawn from the posterior. If
+#'    FALSE (default), the mean curve and associated Credible Interval are
+#'    displayed.
+#' @param nb_samples A number, indicating the number of samples to be drawn from
+#'    the predictive posterior distribution. For two-dimensional graphs, only
+#'    one sample can be displayed.
 #' @param prob_CI A number between 0 and 1 (default is 0.95), indicating the
 #'    level of the Credible Interval associated with the posterior mean curve.
 #'    If this this argument is set to 1, the Credible Interval is not displayed.
@@ -106,7 +113,8 @@ plot_db <- function(data, cluster = FALSE, legend = FALSE) {
 #' @return Visualisation of a Magma or GP prediction (optional: display data
 #'    points, training data points and the prior mean function). For 1-D
 #'    inputs, the prediction is represented as a mean curve and its associated
-#'    95%  Credible Interval, or as a heatmap of probabilities if
+#'    95%  Credible Interval, as a collection of samples drawn from the
+#'    posterior if \code{samples} = TRUE, or as a heatmap of probabilities if
 #'    \code{heatmap} = TRUE. For 2-D inputs, the prediction is represented as a
 #'    heatmap, where each couple of inputs on the x-axis and y-axis are
 #'    associated with a gradient of colours for the posterior mean values,
@@ -125,6 +133,8 @@ plot_gp <- function(pred_gp,
                     prior_mean = NULL,
                     y_grid = NULL,
                     heatmap = FALSE,
+                    samples = FALSE,
+                    nb_samples = 50,
                     prob_CI = 0.95,
                     size_data = 3,
                     size_data_train = 1,
@@ -216,24 +226,34 @@ plot_gp <- function(pred_gp,
 
   ## Display a heatmap if inputs are 2D
   if (ncol(inputs) == 2) {
-    ## Add the 'Index' column if the prediction comes from 'pred_gif()'
-    if (!is.null(index)) {
-      pred <- pred %>% dplyr::mutate("Index" = index)
-    }
 
-    gg <- ggplot2::ggplot() +
-      ggplot2::geom_raster(
-        data = pred,
-        ggplot2::aes_string(
-          x = names(inputs)[1],
-          y = names(inputs)[2],
-          fill = "Mean",
-          alpha = "CI_Width"
-        ),
-        interpolate = TRUE
-      ) +
-      ggplot2::scale_fill_distiller(palette = "RdPu", trans = "reverse") +
-      ggplot2::scale_alpha_continuous(range = c(0.1, 1), trans = "reverse")
+    if (samples){
+      ## Display samples from the posterior
+      gg <- plot_samples(
+        pred = pred_gp,
+        x_input = x_input,
+        nb_samples = nb_samples)
+
+    } else {
+      ## Add the 'Index' column if the prediction comes from 'pred_gif()'
+      if (!is.null(index)) {
+        pred <- pred %>% dplyr::mutate("Index" = index)
+      }
+
+      gg <- ggplot2::ggplot() +
+        ggplot2::geom_raster(
+          data = pred,
+          ggplot2::aes_string(
+            x = names(inputs)[1],
+            y = names(inputs)[2],
+            fill = "Mean",
+            alpha = "CI_Width"
+          ),
+          interpolate = TRUE
+        ) +
+        ggplot2::scale_fill_distiller(palette = "RdPu", trans = "reverse") +
+        ggplot2::scale_alpha_continuous(range = c(0.1, 1), trans = "reverse")
+    }
 
     if (!is.null(data)) {
       ## Round the 'Output' values to reduce size of labels on the graph
@@ -317,6 +337,13 @@ plot_gp <- function(pred_gp,
         ) +
         ggplot2::labs(fill = "Proba CI") +
         ggplot2::ylab("Output")
+    } else if (samples){
+      ## Display samples from the posterior
+      gg <- plot_samples(
+        pred = pred_gp,
+        x_input = x_input,
+        nb_samples = nb_samples)
+
     } else {
       ## Display a classic curve otherwise
       ## Add the 'Index' column if the prediction comes from 'pred_gif()'
@@ -394,79 +421,63 @@ plot_gp <- function(pred_gp,
 #' @export
 plot_magma <- plot_gp
 
-#' Display realisations from a posterior GP
+#' Display realisations from a (mixture of) GP prediction
 #'
-#' Display samples drawn from a posterior GP distribution.
-#' According to the dimension of the inputs, the graph may be a curve or a
-#' heatmap.
+#' Display samples drawn from the posterior of a GP, Magma or
+#' MagmaClust prediction. According to the dimension of the inputs, the graph
+#' may represent curves or a heatmap.
 #'
+#' @param pred A list, typically coming from \code{\link{pred_gp}},
+#'    \code{\link{pred_magma}} or \code{\link{pred_magmaclust}} functions, using
+#'    the argument 'get_full_cov = TRUE'. Required elements: \code{pred},
+#'    \code{cov}. This argument is needed if \code{samples} is missing.
 #' @param samples A tibble or data frame, containing the samples generated from
-#'    a GP prediction. Required columns: \code{Input}, \code{Sample},
-#'    \code{Output}.  This argument is needed if \code{pred_gp} is missing.
-#' @param pred_gp A tibble or data frame, typically coming from
-#'    \code{\link{pred_magma}} or \code{\link{pred_gp}} functions. Required
-#'    columns: \code{Input}, \code{Mean}, \code{Var}. Additional covariate
-#'    columns may be present in case of multi-dimensional inputs. This argument
-#'    is needed if \code{samples} is missing.
+#'    a GP, Magma, or MagmaClust prediction. Required columns: \code{Input},
+#'    \code{Sample}, \code{Output}.  This argument is needed if \code{pred}
+#'    is missing.
 #' @param nb_samples A number, indicating the number of samples to be drawn from
 #'    the predictive posterior distribution. For two-dimensional graphs, only
 #'    one sample can be displayed.
-#' @param x_input A vector of character strings, indicating which input should
-#'    be displayed. If NULL(default) the 'Input' column is used for the x-axis.
-#'    If providing a 2-dimensional vector, the corresponding columns are used
-#'    for the x-axis and y-axis.
-#' @param data (Optional) A tibble or data frame, containing the data used in
-#'    the GP prediction.
-#' @param data_train (Optional) A tibble or data frame, containing the training
-#'    data of the Magma model. The data set should have the same format as the
-#'    \code{data} argument with an additional column 'ID' for identifying the
-#'    different individuals/tasks. If provided, those data are displayed as
-#'    backward colourful points (each colour corresponding to one
-#'    individual/task).
-#' @param prior_mean (Optional) A tibble or a data frame, containing the 'Input'
-#'    and associated 'Output' prior mean parameter of the GP prediction.
-#' @param size_data A number, controlling the size of the \code{data} points.
-#' @param size_data_train A number, controlling the size of the
-#'    \code{data_train} points.
-#' @param alpha_data_train A number, between 0 and 1, controlling transparency
-#'    of the \code{data_train} points.
+#' @param x_input A vector of character strings, indicating which 'column'
+#'    should be displayed in the case of multidimensional inputs. If
+#'    NULL(default) the Input' column is used for the x-axis. If providing a
+#'    2-dimensional vector, the corresponding columns are used for the x-axis
+#'    and the y-axis.
 #'
-#' @return Draw and visualise from a posterior distribution from Magma or GP
-#'    prediction (optional: display data points, training data points and the
-#'    prior mean function).
+#' @return Graph of samples drawn from a posterior distribution of a GP,
+#'    Magma, or MagmaClust prediction.
 #' @export
 #'
 #' @examples
 #' TRUE
-plot_samples <- function(samples = NULL,
-                         pred_gp = NULL,
+plot_samples <- function(pred = NULL,
+                         samples = NULL,
                          nb_samples = 50,
-                         x_input = NULL,
-                         data = NULL,
-                         data_train = NULL,
-                         prior_mean = NULL,
-                         size_data = 3,
-                         size_data_train = 1,
-                         alpha_data_train = 0.5) {
+                         x_input = NULL
+                         ) {
 
   ## Check samples availability or draw samples from the prediction
   if(is.null(samples)){
-    ## Check whether 'pred_gp' exists
-    if(is.null(pred_gp)){
-      stop("Either 'sample' or 'pred_gp' is needed as an argument.")
+    ## Check whether 'pred' exists
+    if(is.null(pred)){
+      stop("Either 'sample' or 'pred' is needed as an argument.")
     }
 
-    ## Check 'pred_gp' format
-    if (is.data.frame(pred_gp) | !is.list(pred_gp)) {
+    ## Check 'pred' format
+    if (!(is.list(pred) & ('cov' %in% names(pred))) ) {
       stop(
-        "The 'pred_gp' argument should be a list containing 'pred' and 'cov' ",
+        "The 'pred' argument should be a list containing 'pred' and 'cov' ",
         "elements. Consider re-running the prediction function using the ",
         "argument 'get_full_cov' = TRUE."
       )
     }
 
-    samples = sample_gp(pred_gp = pred_gp, nb_samples = nb_samples)
-
+    ## Check whether 'pred' is a GP/Magma or a MagmaClust prediction
+    if ('mixture' %in% names(pred)){
+      samples = sample_magmaclust(pred_clust = pred, nb_samples = nb_samples)
+    } else {
+      samples = sample_gp(pred_gp = pred, nb_samples = nb_samples)
+    }
   }
 
   ## Get the inputs that should be used
@@ -495,22 +506,6 @@ plot_samples <- function(samples = NULL,
       ) +
       ggplot2::scale_fill_distiller(palette = "RdPu", trans = "reverse")
 
-    if (!is.null(data)) {
-      ## Round the 'Output' values to reduce size of labels on the graph
-      data <-
-        data %>% dplyr::mutate(Output = round(.data$Output, 1))
-
-      gg <- gg + ggplot2::geom_label(
-        data = data,
-        ggplot2::aes_string(
-          x = names(inputs)[1],
-          y = names(inputs)[2],
-          label = "Output",
-          fill = "Output"
-        ),
-        size = 3
-      )
-    }
   } else if (ncol(inputs) == 1) {
     gg <- ggplot2::ggplot() +
       ggplot2::geom_line(
@@ -523,55 +518,7 @@ plot_samples <- function(samples = NULL,
         color = "#FA9FB5",
         alpha = 0.3
       ) +
-      ggplot2::guides(col = "none")
-
-    ## Display the training data if provided
-    if (!is.null(data_train)) {
-      gg <- gg + ggplot2::geom_point(
-        data = data_train,
-        ggplot2::aes_string(
-          x = names(inputs)[1],
-          y = "Output",
-          col = "ID"
-        ),
-        size = size_data_train,
-        alpha = alpha_data_train
-      ) +
-        ggplot2::guides(color = "none")
-    }
-
-    ## Display the observed data if provided
-    if (!is.null(data)) {
-      gg <- gg + ggplot2::geom_point(
-        data = data,
-        ggplot2::aes_string(
-          x = names(inputs)[1],
-          y = "Output"
-        ),
-        size = size_data,
-        shape = 20
-      )
-    }
-
-    ## Display the (hyper-)prior mean process if provided
-    if (!is.null(prior_mean)) {
-      if (names(inputs)[1] %in% names(prior_mean)) {
-        gg <- gg +
-          ggplot2::geom_line(
-            data = prior_mean,
-            ggplot2::aes_string(x = names(inputs)[1], y = "Output"),
-            linetype = "dashed"
-          )
-      } else {
-        warning(
-          "The ",
-          names(inputs)[1],
-          " column does not exist in the ",
-          "'prior_mean' argument. The mean function cannot be displayed."
-        )
-      }
-    }
-
+      ggplot2::guides(group = "none")
   } else {
     stop(
       "Impossible to display inputs with dimensions greater than 2. Please ",
@@ -751,10 +698,17 @@ plot_gif <- function(pred_gp,
 #'    probabilities should be computed for heatmaps of 1-dimensional
 #'    predictions. If NULL (default), a vector of length 50 is defined, ranging
 #'    between the min and max 'Output' values contained in \code{pred}.
-#' @param heatmap A logical value indicating whether the GP prediction should be
+#' @param heatmap A logical value indicating whether the GP mixture should be
 #'    represented as a heatmap of probabilities for 1-dimensional inputs. If
 #'    FALSE (default), the mean curve (and associated Credible Interval if
 #'    available) are displayed.
+#' @param samples A logical value indicating whether the GP mixture should be
+#'    represented as a collection of samples drawn from the posterior. If
+#'    FALSE (default), the mean curve (and associated Credible Interval if
+#'    available) are displayed.
+#' @param nb_samples A number, indicating the number of samples to be drawn from
+#'    the predictive posterior distribution. For two-dimensional graphs, only
+#'    one sample can be displayed.
 #' @param prob_CI A number between 0 and 1 (default is 0.95), indicating the
 #'    level of the Credible Interval associated with the posterior mean curve.
 #'    If this this argument is set to 1, the Credible Interval is not displayed.
@@ -791,6 +745,8 @@ plot_magmaclust <- function(pred_clust,
                             prior_mean = NULL,
                             y_grid = NULL,
                             heatmap = FALSE,
+                            samples = FALSE,
+                            nb_samples = 50,
                             prob_CI = 0.95,
                             size_data = 3,
                             size_data_train = 1,
@@ -879,7 +835,43 @@ plot_magmaclust <- function(pred_clust,
     }
   }
 
-  if (all_clust) {
+  if(samples) {
+    ## Display samples drawn from the posterior mixture of GPs
+    gg <- plot_samples(
+      pred = pred_clust,
+      nb_samples = nb_samples,
+      x_input = x_input
+      )
+
+    ## Display the observed data if provided
+    if (!is.null(data)) {
+      ## Check dimension of the inputs
+      if (ncol(inputs) == 2) {
+        ## Display labels if 2-D
+        gg <- gg + ggplot2::geom_label(
+          data = data,
+          ggplot2::aes_string(
+            x = names(inputs)[1],
+            y = names(inputs)[2],
+            label = "Output",
+            fill = "Output"
+          ),
+          size = 3
+        )
+      } else if (ncol(inputs) == 1){
+        ## Display points if 1-D
+        gg <- gg + ggplot2::geom_point(
+          data = data,
+          ggplot2::aes_string(x = names(inputs)[1], y = "Output"),
+          size = size_data,
+          shape = 20
+        )
+      }
+    }
+    ## Define the adequate title
+    gtitle <- paste0("Samples from a mixture of GPs")
+
+  } else if (all_clust) {
     ## GP visualisation without Credible Interval
     gg <- plot_gp(
       pred_gp = pred_gp,
@@ -888,13 +880,12 @@ plot_magmaclust <- function(pred_clust,
       y_grid = y_grid,
       heatmap = heatmap,
       prob_CI = 0,
-      size_data = size_data,
-      size_data_train = size_data_train,
-      alpha_data_train = alpha_data_train
+      size_data = size_data
     )
 
     ## Define the adequate title
     gtitle <- paste0("Mixture of GP predictions")
+
   } else {
     ## Classic GP visualisation for cluster-specific predictions
     gg <- plot_gp(
@@ -904,9 +895,7 @@ plot_magmaclust <- function(pred_clust,
       y_grid = y_grid,
       heatmap = heatmap,
       prob_CI = prob_CI,
-      size_data = size_data,
-      size_data_train = size_data_train,
-      alpha_data_train = alpha_data_train
+      size_data = size_data
     )
 
     ## Define the adequate title